+++ aspects/fluid/fluid/GunnsFluidTraceCompounds.cpp
@@ -453,7 +453,7 @@ int GunnsFluidTraceCompounds::find(const ChemicalCompound::Type& type,
 
     /// - Otherwise throw an exception.
     std::ostringstream msg;
-    msg << "An invalid compound type (" << type << ") was specified.";
+    msg << "An invalid compound type (" << static_cast<unsigned int>(type) << ") was specified.";
     TS_HS_EXCEPTION(TS_HS_ERROR, TS_HS_GUNNS, msg.str(),
                     TsOutOfBoundsException, "Input Argument Out of Range", mName);
 }
